# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/activities": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "List Activities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "active page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " count": {
                                            "type": "integer"
                                        },
                                        "Results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Activity"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/activities/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Get Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Activity"
                        }
                    }
                }
            }
        },
        "/api/activities/{id}/gpx": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Export Activity GPX",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Strava Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/activities/{id}/laps": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Get Activity Laps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Strava Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " count": {
                                            "type": "integer"
                                        },
                                        "Results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/strava.LapEffortSummary"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/athletes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "athlete"
                ],
                "summary": "List Athletes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "active page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " count": {
                                            "type": "integer"
                                        },
                                        "Results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Athlete"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/athletes/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "athlete"
                ],
                "summary": "Get Athlete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Athlete ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Athlete"
                        }
                    }
                }
            }
        },
        "/api/auth/authorization-url": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Access Token from Auth Code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Callback URL",
                        "name": "callback_url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Access Token from Auth Code",
                "parameters": [
                    {
                        "description": "Access Token Input",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.getAccessToken.tokenRequestInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/strava.AuthorizationResponse"
                        }
                    }
                }
            }
        },
        "/api/dashboards": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "List Dashboards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "active page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " count": {
                                            "type": "integer"
                                        },
                                        "Results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Dashboard"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Create Dashboard",
                "parameters": [
                    {
                        "description": "Dashboard Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createDashboard.dashboardInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Dashboard"
                        }
                    }
                }
            }
        },
        "/api/dashboards/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get Dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Dashboard"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Update Dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dashboard Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateDashboard.dashboardInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Dashboard"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Delete Dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/dashboards/{id}/components": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "List Dashboard Components",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " count": {
                                            "type": "integer"
                                        },
                                        "Results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Component"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Create Dashboard Components",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Component Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createComponent.componentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Component"
                        }
                    }
                }
            }
        },
        "/api/dashboards/{id}/components/{cpid}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Update Dashboard Component",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Component ID",
                        "name": "cpid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Component Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateComponent.componentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Component"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Delete Dashboard Component",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Component ID",
                        "name": "cpid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/dashboards/{id}/components/{cpid}/run": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Run Dashboard Component",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Component ID",
                        "name": "cpid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/queue.TaskResult"
                        }
                    }
                }
            }
        },
        "/api/dashboards/{id}/run": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Run Dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/queue.TaskResult"
                        }
                    }
                }
            }
        },
        "/api/gears": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gear"
                ],
                "summary": "List Gears",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "active page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " count": {
                                            "type": "integer"
                                        },
                                        "Results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Gear"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/gears/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "gear"
                ],
                "summary": "Get Gear",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gear ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gear"
                        }
                    }
                }
            }
        },
        "/api/user/config": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get Application Config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Upsert Application Config",
                "parameters": [
                    {
                        "description": "Config Input",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                }
            }
        },
...
        "/api/user/query": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Run Query",
                "parameters": [
                    {
                        "description": "Query Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.runQuery.queryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/queue.TaskResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.PaginatedResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "results": {}
            }
        },
...
        "strava.LapEffortSummary": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        }
                    }
                },
                "athlete": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        }
                    }
                },
                "average_cadence": {
                    "type": "number"
                },
                "average_heartrate": {
                    "type": "number"
                },
                "average_speed": {
                    "type": "number"
                },
                "average_watts": {
                    "type": "number"
                },
                "distance": {
                    "type": "number"
                },
                "elapsed_time": {
                    "type": "integer"
                },
                "end_index": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lap_index": {
                    "type": "integer"
                },
                "max_heartrate": {
                    "type": "number"
                },
                "max_speed": {
                    "type": "number"
                },
                "moving_time": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "start_date_local": {
                    "type": "string"
                },
                "start_index": {
                    "type": "integer"
                },
                "total_elevation_gain": {
                    "type": "number"
                }
            }
        },
...
    }
}